fix: 全量修复缺陷报告 12 项缺陷并补齐测试（27/27 全绿）
按 docs/defect_analysis_report_master_20260908.md 完成 S0-S3 全部修复：

S0 构建阻断（DEF-001）：
- 补齐 View 层缺失头文件：Renderer.h / dto.h / view_primitives.h
- 补 src/Controller/Food/CMakeLists.txt（food 库目标）
- 补 tests/{bag,tool,player_serializer,view,weather} 五个缺失子目录
- master 恢复可 configure + build + ctest

S1 严重缺陷：
- DEF-003: mine.start/fish.tick 交互等待循环移出 worldMutex，
  改为分段短暂加锁（校验/产出/状态还原各自持锁），后台时间线程不再饿死
- DEF-004: time.scale 缺参/非法参数经 opt_double 兜底回退当前倍率，
  杜绝 std::out_of_range 未捕获崩溃

S2 中等缺陷：
- DEF-005: weather_controller RNG 补 thread_local
- DEF-006: Weather 查表改 config_for(type) 按 type 匹配；
  ToolController 数组改 std::map<ToolId, Tool>，消除枚举序-数组序隐式耦合
- DEF-007: 金币链路全程 64 位（Market buy/sell、PlayerSerializer、
  main gold、DTO MarketView/BlacksmithView），大额交易不再截断
- DEF-008: FarmLand::fertilize 改为按「周期/speedUp」直接补贴进度，
  播种后立即施肥有效，补贴钳制到周期上限
- DEF-009: fish.tick 饱食度扣减钳制非负

S3 低危：
- DEF-011: Move::MoveTo 改收 unique_ptr<Position>，编译期强制所有权
- DEF-012: 全局 rand 收敛到 thread_local mt19937
  （weather/FishingController/Market/main），移除 srand
- DEF-013: Punpkin.h 重命名为 Pumpkin.h 并同步 include
- DEF-014: build_mine_ctx 档位显式化，0/1 级归最低速 Core
- DEF-015: 作物/鱼/商店原型对象改 static 存储期并注释所有权

新增测试（按模块分功能，8 个目标 34 个用例）：
- tests/bag: copy_semantics（DEF-002 回归）+ stacking
- tests/tool: lookup（DEF-006 回归）
- tests/weather: config_lookup（DEF-006 回归）
- tests/view: render（S0 头文件回归 + DTO 组装）
- tests/player_serializer: roundtrip（DEF-002/007 回归）
- tests/farm: fertilize（DEF-008 回归）+ water_grow

验证：Ninja 全量构建 301/301 无警告；ctest 27/27 通过；
graphify 图谱已重建（1813 节点/2592 边/121 社区）

+++ src/Tool/src/PlayerSerializer.cpp
@@ -8,7 +8,7 @@
 #include <sstream>
 
 bool PlayerSerializer::Save(const std::string& filename, const Player& player, const Game& game,
-                            int gold, const mud::tool::ToolController& tools,
+                            long long gold, const mud::tool::ToolController& tools,
                             std::int64_t totalGameMinutes) {
     std::ofstream file(filename);
     if (!file.is_open()) {
@@ -66,7 +66,7 @@ bool PlayerSerializer::Save(const std::string& filename, const Player& player) {
     return Save(filename, player, game, 0, mud::tool::ToolController{}, 0);
 }
 
-bool PlayerSerializer::Load(const std::string& filename, Player& player, Game& game, int& gold,
+bool PlayerSerializer::Load(const std::string& filename, Player& player, Game& game, long long& gold,
                             mud::tool::ToolController& tools, std::int64_t& totalGameMinutes) {
     std::ifstream file(filename);
     if (!file.is_open()) {
@@ -77,10 +77,9 @@ bool PlayerSerializer::Load(const std::string& filename, Player& player, Game& g
     StateCode state = Waiting;
     int satiety = 100, maxSatiety = 100;
     int farmingExp = 0, fishExp = 0, mineExp = 0;
-    int goldInFile = 0;
+    long long goldInFile = 0;   // DEF-007：金币 64 位，支持大额数值
     bool hasGold = false;
     bool hasTotalMinutes = false;
-
     long long saveOpenYear = 0, saveOpenMonth = 1, saveOpenDay = 0, saveOpenHour = 0, saveOpenMinute = 0;
     long long totalPlaySeconds = 0;
 
@@ -113,7 +112,7 @@ bool PlayerSerializer::Load(const std::string& filename, Player& player, Game& g
             } else if (key == "mineExp") {
                 mineExp = std::stoi(value);
             } else if (key == "gold") {
-                goldInFile = std::stoi(value);
+                goldInFile = std::stoll(value);
                 hasGold = true;
             } else if (key == "saveOpenYear") {
                 saveOpenYear = std::stoll(value);
@@ -207,7 +206,7 @@ bool PlayerSerializer::Load(const std::string& filename, Player& player, Game& g
 
 bool PlayerSerializer::Load(const std::string& filename, Player& player) {
     Game game;
-    int gold = 0;
+    long long gold = 0;
     mud::tool::ToolController tools;
     std::int64_t total = -1;
     return Load(filename, player, game, gold, tools, total);
